% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfMintBubblesTransactions}
\alias{dfMintBubblesTransactions}
\title{Mint bubbles transations data}
\format{A data frame with 384 rows and 6 columns.}
\source{
\url{github.com}
}
\usage{
dfMintBubblesTransactions
}
\description{
Some person's banking financial data.
}
\keyword{datasets}
