% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObfuscationFuncitons.R
\name{ObfuscateAmounts}
\alias{ObfuscateAmounts}
\title{Amounts obfuscation.}
\usage{
ObfuscateAmounts(amounts, randomizationFraction = 0.1, factor = NULL)
}
\arguments{
\item{amounts}{Amounts vector to be randomized.}

\item{randomizationFraction}{The fraction to randomize with.}

\item{factor}{A factor to multiply the randomized fractions with.}
}
\value{
A numerical vector.
}
\description{
Changes amounts according to their values.
Same amounts at different records would have the same obfuscated values.
}
\details{
All unique amounts are mapped into unique amounts.
The original motivation for this function was to obfuscate money amounts.
}
