% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObfuscationFuncitons.R
\name{ObfuscateDescriptions}
\alias{ObfuscateDescriptions}
\title{Descriptions obfuscation.}
\usage{
ObfuscateDescriptions(descriptions, mappedWords = NULL,
  preservedWords = NULL, randomWordsQ = TRUE,
  randomWordLengthsQ = TRUE)
}
\arguments{
\item{descriptions}{A character vector.}

\item{mappedWords}{A character vector with named elements; names are mapped to the elements.}

\item{preservedWords}{A character list/vector of words that should be kept the same.}

\item{randomWordsQ}{Should the words that are not in \code{mappedWords} and \code{preservedWords} be randomly mapped?}

\item{randomWordLengthsQ}{Should the random words have random lengths?}
}
\value{
A character vector with obfuscated descriptions.
}
\description{
Obfuscates transactions descriptions with specified word mapping,
preserved words. All other words are mapped to random words with corresponding lengths
or random lengths.
}
