% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObfuscationFuncitons.R
\name{ShuffleAmountsByColumnValuesPartition}
\alias{ShuffleAmountsByColumnValuesPartition}
\title{Shuffle amounts by column partition.}
\usage{
ShuffleAmountsByColumnValuesPartition(data, amountsColNames,
  splitColNames = NULL)
}
\arguments{
\item{data}{A data frame of banking transactions.}

\item{amountsColNames}{A list of column names with amounts to be shuffled.}

\item{splitColNames}{NULL or a list of column names for splitting the transactions (for \code{\link{ddply}}.)}
}
\value{
A data frame.
}
\description{
Shuffles one or several columns within subsets of the transactions derived by unique
values of specified columns.
}
\details{
The arguments allow the shuffling of several amount columns with the same permutation.
Note that the original order of the transactions data frame is lost.
}
