% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObfuscationFuncitons.R
\name{ShuffleAmountsInRegexSubsets}
\alias{ShuffleAmountsInRegexSubsets}
\title{Shuffle amounts by regex subsets.}
\usage{
ShuffleAmountsInRegexSubsets(data, amountsColNames, forPatternsColName,
  patterns)
}
\arguments{
\item{data}{A data frame of banking transactions.}

\item{amountsColNames}{A list of column names with amounts to be shuffled.}

\item{forPatternsColName}{NULL or a column name to the values of which the patterns are applied.}

\item{patterns}{regex patterns to be used do derive subsets from the values of forPatternsColName}
}
\description{
Shuffles one or several columns within subsets of transactions derived by
by adherence to regex patterns.
}
\details{
The arguments allow the shuffling of several amount columns with the same permutation.
}
